refactor(portcullis): extract flow_graph test module to flow_graph_tests.rs
Move 2182-line inline test module from flow_graph.rs to separate file.
flow_graph.rs: 3451 → 1270 lines (2181 line reduction).

All 86 flow_graph unit tests + 14 kernel red team tests pass.
